7. There are only 3 colors, 10 digits, and 7 notes; it's what we do with them that's important

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -4,7 +4,7 @@
  * @i : an integer whose last digit is to be printed.
  * Return: the value of the last digit.
  */
-int print_last_digit(int i)
+int print_last_digit(long i)
 {
 	if (i < 0)
 		i *= -1;
